purchase with coupon

+++ SWP391_Backend/src/main/java/com/entity/OrderEntity.java
@@ -40,7 +40,7 @@ public class OrderEntity {
     Instant expiredAt;
 
     @Column(name = "total_amount")
-    Integer totalAmount;
+    Double totalAmount;
 
     @ManyToOne
     @JoinColumn(name = "user_id")
